Split Index Builder into multiple classes
